requirements.txt
playwright
python-socketio
aiohttp
tabulate